telegram-bot/requirements.txt
python-telegram-bot
aiohttp
redis
python-dotenv